Use a dynamically linked version of busybox to save space
Needed the libm library and needed to modify 8.sh and also
umount-live to deal with the dynamic libs.

8.sh needed to change because it relied on top level symlinks to
/live/linux and /live/aufs which had to go because they broke the
dyanamic busybox.

The umount-live program had to change because when it is called from
the modified init.d shutdown scripts it is linked to libs in the
aufs.  So we call ourselves again after the pivot root and this gives
us the live libs so we can cleanly umount the aufs and all that it
depends on.